+++ interpreter/extern-module-definition.go
@@ -1,0 +1,5 @@
+package interpreter
+
+type ExternalDefinition interface {
+	Lookup(name string, env *Environment) (RuntimeValue, bool)
+}